backend/foodgram/requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
Django==2.2.19
django-filter==21.1
djangorestframework==3.12.4
djoser==2.1.0
drf_extra_fields==3.4.0
flake8==4.0.1
gunicorn==20.0.4
idna==3.3
importlib-metadata==1.7.0
iniconfig==1.1.1
packaging==21.3
pluggy==0.13.1
Pillow==9.1.1
psycopg2-binary==2.8.6
py==1.11.0
pyparsing==3.0.7
python-dotenv==0.20.0
pytz==2022.1
reportlab==3.6.10
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
urllib3==1.26.9
webcolors==1.12
zipp==3.8.0